+++ test/emu_mps/test_hamiltonian.py
@@ -565,64 +565,3 @@ def test_truncation_nn(hamiltonian_type):
         sv,
         expected,
     )
-
-
-@pytest.mark.parametrize(
-    "hamiltonian_type",
-    [
-        HamiltonianType.Rydberg,
-        HamiltonianType.XY,
-    ],
-)
-@pytest.mark.parametrize(
-    "n_qubits",
-    [10, 25, 50],
-)
-def test_interaction_matrix_shape(hamiltonian_type, n_qubits):
-    def test_assert_make_H(int_mat: int) -> None:
-        with pytest.raises(
-            AssertionError, match="Interaction matrix is not symmetric and zero diag"
-        ):
-            make_H(
-                interaction_matrix=int_mat,
-                hamiltonian_type=hamiltonian_type,
-                num_gpus_to_use=0,
-            )
-
-    wrong_interaction_matrix = torch.tensor([[]])  # empty matrix
-    test_assert_make_H(wrong_interaction_matrix)
-
-    wrong_interaction_matrix = torch.randn(n_qubits, dtype=torch.float64)  # 1D not matrix
-    test_assert_make_H(wrong_interaction_matrix)
-
-    wrong_interaction_matrix = torch.randn(
-        n_qubits, n_qubits, n_qubits, dtype=torch.float64
-    )  # 3D not matrix
-    test_assert_make_H(wrong_interaction_matrix)
-
-    wrong_interaction_matrix = torch.randn(
-        n_qubits, n_qubits + 1, dtype=torch.float64
-    )  # not square
-    test_assert_make_H(wrong_interaction_matrix)
-
-    wrong_interaction_matrix = torch.randn(
-        n_qubits, n_qubits - 1, dtype=torch.float64
-    )  # not square
-    test_assert_make_H(wrong_interaction_matrix)
-
-    wrong_interaction_matrix = torch.randn(
-        n_qubits, n_qubits, dtype=torch.float64
-    )  # not symmetric
-    test_assert_make_H(wrong_interaction_matrix)
-
-    wrong_interaction_matrix = (
-        wrong_interaction_matrix + wrong_interaction_matrix.T
-    )  # diagonal not 0
-    test_assert_make_H(wrong_interaction_matrix)
-
-    correct_interaction_matrix = wrong_interaction_matrix.fill_diagonal_(0)
-    make_H(
-        interaction_matrix=correct_interaction_matrix,
-        hamiltonian_type=hamiltonian_type,
-        num_gpus_to_use=0,
-    )
